src/discordapp/requirements.txt
discord.py
python-dotenv
boto3